requirements.txt
flask
requests
pymongo
dnspython
python-dotenv
flask-cors